6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,11 +6,11 @@
 */
 int main(void)
 {
-printf("size of char: %llu bytes\n", sizeof(char));
-printf("size of float: %llu bytes\n", sizeof(float));
-printf("size of double: %llu bytes\n" sizeof(double));
-printf("size of int: %llu bytes\n" sizeof(int));
-printf("size of long: %llu bytes\n" sizeof(long));
-printf("size of long long:%llu bytes\n" sizeof(long long));
+printf("size of char: %llu bytes\n", (unsigned long) sizeof(char));
+printf("size of float: %llu bytes\n", (unsigned long) sizeof(float));
+printf("size of double: %llu bytes\n", (unsigned long) sizeof(double));
+printf("size of int: %llu bytes\n", (unsigned long)  sizeof(int));
+printf("size of long: %llu bytes\n", (unsigned long) sizeof(long));
+printf("size of long long:%llu bytes\n", (unsigned long) sizeof(long long));
 return (0);
 }
